Update common.sh to read filepath_map_json to generate file locations
These changes make this reliant on a earlier change to the web ui that includes filepath_map.json

This change also rounds off making all 3 languages (PHP, Python & Bash) use a single source to generate their file paths

+++ scripts/update_birdnet_snippets.sh
@@ -248,8 +248,9 @@ if grep -q '^MODEL=BirdNET_GLOBAL_3K_V2.2_Model_FP16$' "$etc_birdnet_conf_path";
   sudo chmod +x "$SCRIPTS_DIR"/install_language_label_nm.sh && "$SCRIPTS_DIR"/install_language_label_nm.sh -l "$language"
 fi
 
-# Link in new bash script common.sh to local/bin
-ln -sf ${SCRIPTS_DIR}/common.sh /usr/local/bin/
+# Link in new bash script common.sh to local/bin and make sure the config directory is also
+[ -L /usr/local/bin/common.sh ] || ln -sf ${SCRIPTS_DIR}/common.sh /usr/local/bin/
+[ -L /usr/local/bin/config ] || ln -sf ${BIRDNET_PI_DIR}/config /usr/local/bin/
 
 
 sudo systemctl daemon-reload
